chore: rename to getUrlRegex.ts
