+++ react-components/src/utilities/getUrlRegex.ts
